requirements.txt
beautifulsoup4==4.12.3
Flask==3.1.0
numpy==2.2.2
pandas==2.2.3
python-dotenv==1.0.1
selenium==4.28.1
tqdm==4.67.1
